Make the configuration adhere to documentation
This fucks up intelligent python autocompletion suggestion,
however, now, the configuaration objects (from config
import xyz_conf) are constructed according to the content
of the docstring of the module. This means changing the
configuration is as easy as writting down the name of the
option and explaining what it does (former is optional).